Remove support for legacy/unsupported packets
The way to tell the client that a packet is not supported is by
replying the empty string, not returning an error.  In some cases,
e-server gets that wrong.

Also, there's no real point in maintaining code that recognizes the
packet and then does nothing with it.  It just makes the stub harder
to read and maintain.

Also, GdbServer::rspSet deletes the packet object, which is wrong,
since nowadays the packet object is reused for all interactions with
the client.