e-server: Treat exceptions as the core has halted.
The e-server  gets stuck on software exceptions when the software exception
interrupt is masked.

When a software exception occurs, the core asserts ILAT[1] to trigger the
software exception ISR. However, when the software exception interrupt is
masked, the core will just get stuck on the instruction that triggered the
exception. But it will not set DEBUGSTATUS[1], so e-server will still think
the core is running.

This patch adds a check to Thread::isHalted() for exceptions. If an
exception is detected, halt () will be called to 'really' halt the core.

N.B: This patch doesn't check whether the software exception interrupt is
masked, nor if there is a valid branch to an exception ISR in IVT[1].

e-server/
	* src/Thread.cpp (Thread::isHalted): Check for exceptions
	and treat them as the core has halted.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>

Conflicts:
	e-server/ChangeLog