e-server: Don't cache DEBUGSTATE
Disable caching of DEBUGSTATE so e-server can sit in the background and
we don't have to start a new e-server between every program run.

Until there is a way for the Kernel driver (or e-hal + PAL) to report a
chip reset (from another program) to e-server, we must not assume that
we know this value (a chip reset clears DEBUGSTATUS) beforehand.
Reporting a stale value here can lead to a read from a general-purpose
register, while the core is not halted, elsewhere in e-server. Such
reads have priority over the core's pipeline, and will corrupt core
state as register writes in the E2 (?) stage will be discarded when
there is a conflict. Note that we still have a race in here if a reset
is issued between a call to isHalted and a call to readReg, but this is
the best we can do atm.

e-server/
	* src/Thread.cpp (Thread::isHalted): Never use cached mDebugState
	member value. Add comment explaining why.
	(Thread::halt): Use isHalted to get and set value of mDebugState.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>